// Public/src/models/ModelType.Model.ts
export interface IModelTypeModel {
    Active: boolean
    BrandMasterId: number
    Created: string
    CreatedBy: number
    BrandMasterName: string
    ProductTypeName: string
    ProductTypeEnumName: string
    DateOfIntroduction: any
    DisplayInList: boolean
    DisplayName: string
    EnumName: string
    Id: number
    IsValid: boolean
    Modified: string
    ModifiedBy: number
    Name: string
    RowOrder: number
    ThumbnailPath: any
    TotalRowsCount: any
    ValidationErrors: any
}